WIP: change mapping of uint{8,16,32,64}
This changes the mapping of uint types to non-widening yang.common
classes and deals with them being used by yang.{data,model}.api.

JIRA: YANGTOOLS-1017
JIRA: MDSAL-330

+++ binding/mdsal-binding-generator-impl/src/test/java/org/opendaylight/mdsal/binding/generator/impl/ChoiceCaseGenTypesTest.java
@@ -121,7 +121,7 @@ public class ChoiceCaseGenTypesTest extends AbstractTypesTest {
 
         genType = checkGeneratedType(genTypes, "AnonymousLock", pcgPref
                 + ".netconf.state.datastores.datastore.locks.lock.type"); // choice
-        containsMethods(genType, new NameTypePattern("getLockTime", "Long"));
+        containsMethods(genType, new NameTypePattern("getLockTime", "Uint32"));
         containsInterface("LockType", genType);
 
         genType = checkGeneratedType(genTypes, "LeafAugCase", pcgPref
